test_backends: factor out 'TestMixed' class
