+++ tests/test_backends/test_tensorflow.py
@@ -3,7 +3,6 @@ import unittest
 
 import numpy as np
 import numpy.random as rnd
-import numpy.testing as np_testing
 import tensorflow as tf
 
 from pymanopt.function import TensorFlow
@@ -107,19 +106,20 @@ class TestTensor3(_backend_tests.TestTensor3):
         self.cost = cost
 
 
-class TestMixed(unittest.TestCase):
-    # Test autograd on a tuple containing vector, matrix and tensor3.
+class TestMixed(_backend_tests.TestMixed):
     def setUp(self):
-        n1 = self.n1 = 3
-        n2 = self.n2 = 4
-        n3 = self.n3 = 5
-        n4 = self.n4 = 6
-        n5 = self.n5 = 7
-        n6 = self.n6 = 8
+        super().setUp()
 
-        x = tf.Variable(tf.zeros([n1]))
-        y = tf.Variable(tf.zeros([n2, n3]))
-        z = tf.Variable(tf.zeros([n4, n5, n6]))
+        n1 = self.n1
+        n2 = self.n2
+        n3 = self.n3
+        n4 = self.n4
+        n5 = self.n5
+        n6 = self.n6
+
+        x = tf.Variable(tf.zeros(n1, dtype=np.float64))
+        y = tf.Variable(tf.zeros([n2, n3], dtype=np.float64))
+        z = tf.Variable(tf.zeros([n4, n5, n6], dtype=np.float64))
 
         @TensorFlow(x, y, z)
         def cost(x, y, z):
@@ -128,87 +128,6 @@ class TestMixed(unittest.TestCase):
                     tf.exp(tf.reduce_sum(z ** 2)))
 
         self.cost = cost
-        self.arg = [x, y, z]
-
-        self.y = y = (rnd.randn(n1).astype(np.float32) * 1e-3,
-                      rnd.randn(n2, n3).astype(np.float32) * 1e-3,
-                      rnd.randn(n4, n5, n6).astype(np.float32) * 1e-3)
-        self.a = a = (rnd.randn(n1).astype(np.float32) * 1e-3,
-                      rnd.randn(n2, n3).astype(np.float32) * 1e-3,
-                      rnd.randn(n4, n5, n6).astype(np.float32) * 1e-3)
-
-        self.correct_cost = (np.exp(np.sum(y[0]**2)) +
-                             np.exp(np.sum(y[1]**2)) +
-                             np.exp(np.sum(y[2]**2)))
-
-        # CALCULATE CORRECT GRAD
-        g1 = 2 * y[0] * np.exp(np.sum(y[0] ** 2))
-        g2 = 2 * y[1] * np.exp(np.sum(y[1] ** 2))
-        g3 = 2 * y[2] * np.exp(np.sum(y[2] ** 2))
-
-        self.correct_grad = (g1, g2, g3)
-
-        # CALCULATE CORRECT HESS
-        # 1. VECTOR
-        Ymat = np.matrix(y[0])
-        Amat = np.matrix(a[0])
-
-        diag = np.eye(n1)
-
-        H = np.exp(np.sum(y[0] ** 2)) * (4 * Ymat.T.dot(Ymat) + 2 * diag)
-
-        # Then 'left multiply' H by A
-        h1 = np.array(Amat.dot(H)).flatten()
-
-        # 2. MATRIX
-        # First form hessian tensor H (4th order)
-        Y1 = y[1].reshape(n2, n3, 1, 1)
-        Y2 = y[1].reshape(1, 1, n2, n3)
-
-        # Create an m x n x m x n array with diag[i,j,k,l] == 1 iff
-        # (i == k and j == l), this is a 'diagonal' tensor.
-        diag = np.eye(n2 * n3).reshape(n2, n3, n2, n3)
-
-        H = np.exp(np.sum(y[1] ** 2)) * (4 * Y1 * Y2 + 2 * diag)
-
-        # Then 'right multiply' H by A
-        Atensor = a[1].reshape(1, 1, n2, n3)
-
-        h2 = np.sum(H * Atensor, axis=(2, 3))
-
-        # 3. Tensor3
-        # First form hessian tensor H (6th order)
-        Y1 = y[2].reshape(n4, n5, n6, 1, 1, 1)
-        Y2 = y[2].reshape(1, 1, 1, n4, n5, n6)
-
-        # Create an n1 x n2 x n3 x n1 x n2 x n3 diagonal tensor
-        diag = np.eye(n4 * n5 * n6).reshape(n4, n5, n6, n4, n5, n6)
-
-        H = np.exp(np.sum(y[2] ** 2)) * (4 * Y1 * Y2 + 2 * diag)
-
-        # Then 'right multiply' H by A
-        Atensor = a[2].reshape(1, 1, 1, n4, n5, n6)
-
-        h3 = np.sum(H * Atensor, axis=(3, 4, 5))
-
-        self.correct_hess = (h1, h2, h3)
-
-    def test_compile(self):
-        np_testing.assert_allclose(self.correct_cost, self.cost(self.y))
-
-    def test_grad(self):
-        grad = self.cost.compute_gradient()
-        for k in range(len(grad(self.y))):
-            np_testing.assert_allclose(self.correct_grad[k], grad(self.y)[k],
-                                       rtol=1e-4)
-
-    def test_hessian(self):
-        hess = self.cost.compute_hessian()
-
-        # Now test hess
-        for k in range(len(hess(self.y, self.a))):
-            np_testing.assert_allclose(self.correct_hess[k],
-                                       hess(self.y, self.a)[k], rtol=1e-4)
 
 
 class TestUserProvidedSession(unittest.TestCase):
